Fix get bids info
FIX: get_bids_info()
The gear calls get_bids_info() to get the BIDS Curation report data.  If it passes in a session ID, only info for that session is returned.  Now it passes an empty string if session_only is False (if it is running on something more than just the session).
ENH: 1.0.1 of the bids-client
This also uses version 1.0.1 of the bids-client which has an improved ReproIn template (subject names have special characters removed and MEGRE/MESE acquisitions are handled by anat_file)